+++ lib/src/handlers/handler_interface.dart
@@ -4,8 +4,7 @@ import 'package:mediasoup_client_flutter/src/sctp_parameters.dart';
 import 'package:mediasoup_client_flutter/src/transport.dart';
 import 'package:mediasoup_client_flutter/src/common/enhanced_event_emitter.dart';
 import 'package:mediasoup_client_flutter/src/rtp_parameters.dart';
-import 'package:mediasoup_client_flutter/src/handlers/native.dart'
-  if (dart.library.html) 'package:mediasoup_client_flutter/src/handlers/browser.dart';
+import 'package:mediasoup_client_flutter/src/handlers/unified_plan.dart';
 
 class SCTP_NUM_STREAMS {
   static const int OS = 1024;
@@ -227,7 +226,7 @@ class SetRtpEncodingParametersOptions {
 abstract class HandlerInterface extends EnhancedEventEmitter {
   HandlerInterface() : super();
 
-  static HandlerInterface handlerFactory() => Handler();
+  static HandlerInterface handlerFactory() => UnifiedPlan();
 
   ///@emits @connect - (
   ///    { dtlsParameters: DtlsParameters },